test(core): Move new execution repository tests to @n8n/db (no-changelog)
